requirements.txt
wheel==0.45.1
requests>=2.31.0
pandas>=2.1.0
numpy==1.26.3
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pytest==8.0.0
pytest-cov==4.1.0
pdfplumber>=0.10.3
playwright>=1.41.2
